Add views and crud controller functions for job posts
