+++ app/views/job_posts/edit.html.erb
@@ -1,0 +1,1 @@
+<%= render 'form', job_post: @job_post %>
