+++ app/views/job_posts/index.html.erb
@@ -1,0 +1,9 @@
+<h2>Job Posts</h2>
+  <a href="<%= edit_job_post_path(job_post) %>" class="btn btn-secondary">Edit</a>
+<% if @job_posts.empty? %>
+  <p>No job posts available yet</p>
+<% else %>
+  <% @job_posts.each do |job_post| %>
+    <h2><%= link_to job_post.title, job_post_path(job_post) %></h2>
+  <% end %>
+<% end %>
