+++ app/views/job_posts/show.html.erb
@@ -1,0 +1,1 @@
+<h1><%= @job_posts.title %></h1>
